Address review: deduplicate copy_dir_recursive, use libc::ELOOP, fix clippy
- Extract shared `copy_dir_recursive` and `is_eloop` into `src/util.rs`
- Remove duplicate implementations from workspace.rs, library/mod.rs, api/library.rs
- Replace hardcoded ELOOP errno 40 with `libc::ELOOP` (40 on Linux, 62 on macOS)
- Fix clippy `io_other_error` lint (use `std::io::Error::other`)
